requirements.txt
# ============================================================================
#  Knowledge Graph RAG - Python Dependencies (OpenAI + Graphiti)
# ============================================================================

# Zep Graphiti - Temporal knowledge graph & RAG framework
graphiti-core>=0.27.0

# Additional utilities
python-dotenv>=1.0.0
pandas>=2.1.0
faker>=22.0.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Optional: API support (for future enhancements)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
